Moved into api folder + listing Sketchbook
